requirements.txt
astroid==2.4.2
cachelib==0.1.1
click==7.1.2
colorama==0.4.3
Flask==1.1.2
Flask-Cors==3.0.9
flask-swagger-ui==3.25.0
install==1.3.3
isort==5.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.7.1
mccabe==0.6.1
pylint==2.6.0
pylint-flask==0.6
pylint-plugin-utils==0.6
python-dotenv==0.14.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
uuid==1.30
Werkzeug==1.0.1
wrapt==1.12.1
